Move InvalidEmailAddressFormatException to the common exceptions package.
