+++ src/main/kotlin/hu/netsurf/erp/common/exception/InvalidEmailAddressFormatException.kt
@@ -1,3 +1,3 @@
-package hu.netsurf.erp.usermanagement.exception
+package hu.netsurf.erp.common.exception
 
 class InvalidEmailAddressFormatException : Exception("Nem megfelelő e-mail cím formátum.")
